+++ core/src/com/pacman/model/Pacman.java
@@ -2,7 +2,7 @@ package com.pacman.model;
 
 public class Pacman extends MoveableElement {
 	
-	public Pacman(float x, float y, World w) {
+	public Pacman(int x, int y, World w) {
 		super(x, y, w);
 		setState(State.LEFT);
 		//setLastState(State.LEFT);
