+++ core/src/com/pacman/model/ghosts/GhostPink.java
@@ -4,7 +4,7 @@ import com.pacman.model.World;
 
 public class GhostPink extends Ghost {
 
-	public GhostPink(float x, float y, World w) {
+	public GhostPink(int x, int y, World w) {
 		super(x, y, w);
 		setSpeed(1f);
 	}
